+++ tests/compile-fail/ClrNameAbstractInterface.kt
@@ -1,0 +1,5 @@
+interface ClrNameAbstractInterface {
+    @Suppress("INAPPLICABLE_JVM_NAME")
+    @kotlin.jvm.JvmName("renamed")
+    fun member(): Int
+}